% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.iter.paramtest}
\alias{n.iter.paramtest}
\title{Return the number of iterations performed by a parameter test.}
\usage{
\method{n.iter}{paramtest}(test)
}
\arguments{
\item{test}{An object of type 'paramtest'.}
}
\value{
Returns the number of iterations done in each test.
}
\description{
\code{n.iter.paramtest} extracts information about the number of iterations
(per specific test) performed by a parameter test.
}

