% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.iter}
\alias{n.iter}
\title{Return the number of iterations performed by a parameter test.}
\usage{
n.iter(test, ...)
}
\arguments{
\item{test}{A paramtest object.}

\item{...}{Additional arguments to be passed to the particular method for the
given object tye.}
}
\value{
The form of the value returned by \code{n.iter} depends on the class
  of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{n.iter} is a generic function that extracts information about the
number of iterations (per specific test) performed by a parameter test.
}
\seealso{
\code{\link{n.iter.paramtest}}
}

