% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.paramtest_summary}
\alias{print.paramtest_summary}
\title{Print summary of parameter tests.}
\usage{
\method{print}{paramtest_summary}(x, ...)
}
\arguments{
\item{x}{An object of class 'paramtest_summary', from
\code{\link{summary.paramtest}}.}

\item{...}{Not currently implemented; used to ensure consistency with S3 generic.}
}
\value{
Returns a data frame with one row per set of unique tests.
}
\description{
\code{print.paramtest_summary} prints a summary of the various combinations
of parameter values tested in a given parameter test.
}
\seealso{
\code{\link{summary.paramtest}}
}
