% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{results.paramtest}
\alias{results.paramtest}
\title{Return results of a parameter test.}
\usage{
\method{results}{paramtest}(test)
}
\arguments{
\item{test}{An object of type 'paramtest'.}
}
\value{
Returns a data frame with all the data returned from each test.
}
\description{
\code{results.paramtest} returns the raw data from a parameter test.
}

