% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{results}
\alias{results}
\alias{results.paramtest}
\title{Return results of a parameter test.}
\usage{
results(test, ...)

\method{results}{paramtest}(test, ...)
}
\arguments{
\item{test}{An object of type 'paramtest'.}

\item{...}{Not currently implemented; used to ensure consistency with S3 generic.}
}
\value{
Returns a data frame with all the data returned from each test.
}
\description{
\code{results} returns the raw data from a parameter test.
}
\section{Methods (by class)}{
\itemize{
\item \code{paramtest}: Results for a parameter test.
}}

