% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests.paramtest}
\alias{tests.paramtest}
\title{Return the parameter values that were tested by paramtest.}
\usage{
\method{tests}{paramtest}(test)
}
\arguments{
\item{test}{An object of type 'paramtest'.}
}
\value{
Returns a data frame with one row for each set of tests that
  was performed.
}
\description{
\code{tests.paramtest} extracts information about the set of specific tests
(parameter values) for a parameter test.
}

