% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests}
\alias{tests}
\title{Return the parameter values that were tested by paramtest.}
\usage{
tests(sim, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to the particular method for the
given object tye.}

\item{test}{A paramtest object.}
}
\value{
The form of the value returned by \code{tests} depends on the class
  of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{tests} is a generic function that extracts information about the set of
specific tests (parameter values) for a parameter test.
}
\seealso{
\code{\link{tests.paramtest}}
}

