% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timing.paramtest}
\alias{timing.paramtest}
\title{Return the timing information of a parameter test.}
\usage{
\method{timing}{paramtest}(test)
}
\arguments{
\item{test}{An object of type 'paramtest'.}
}
\value{
Returns an object of class "proc_time" with information about how
  long the parameter test process took.
}
\description{
\code{timing.paramtest} returns the raw data from a parameter test.
}

