% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timing}
\alias{timing}
\alias{timing.paramtest}
\title{Return the timing information of a parameter test.}
\usage{
timing(test, ...)

\method{timing}{paramtest}(test, ...)
}
\arguments{
\item{test}{An object of type 'paramtest'.}

\item{...}{Not currently implemented; used to ensure consistency with S3 generic.}
}
\value{
Returns an object of class "proc_time" with information about how
  long the parameter test process took.
}
\description{
\code{timing} returns the information about how long a parameter test took.
}
\section{Methods (by class)}{
\itemize{
\item \code{paramtest}: Timing information for a paramter test.
}}

