% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.iter}
\alias{n.iter}
\alias{n.iter.paramtest}
\title{Return the number of iterations performed by a parameter test.}
\usage{
n.iter(test, ...)

\method{n.iter}{paramtest}(test, ...)
}
\arguments{
\item{test}{An object of type 'paramtest'.}

\item{...}{Not currently implemented; used to ensure consistency with S3 generic.}
}
\value{
Returns the number of iterations done in each test.
}
\description{
\code{n.iter} extracts information about the number of iterations (per
specific test) performed by a parameter test.
}
\section{Methods (by class)}{
\itemize{
\item \code{paramtest}: Number of iterations for a parameter test.
}}

