% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests}
\alias{tests}
\alias{tests.paramtest}
\title{Return the parameter values that were tested by paramtest.}
\usage{
tests(test, ...)

\method{tests}{paramtest}(test, ...)
}
\arguments{
\item{test}{An object of type 'paramtest'.}

\item{...}{Not currently implemented; used to ensure consistency with S3 generic.}
}
\value{
Returns a data frame with one row for each set of tests that
  was performed.
}
\description{
\code{tests} extracts information about the set of specific tests (parameter
values) for a parameter test.
}
\section{Methods (by class)}{
\itemize{
\item \code{paramtest}: Parameter values for a parameter test.
}}

