% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timing}
\alias{timing}
\title{Return the timing information of a parameter test.}
\usage{
timing(test, ...)
}
\arguments{
\item{test}{A paramtest object.}

\item{...}{Additional arguments to be passed to the particular method for the
given object tye.}
}
\value{
The form of the value returned by \code{timing} depends on the class
  of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{timing} is a generic function that extracts the raw data from a
parameter test.
}
\seealso{
\code{\link{timing.paramtest}}
}

